+++ Others/Kotlin/kotlin-koans/src/ii_collections/n18Sort.kt
@@ -1,0 +1,10 @@
+package ii_collections
+
+fun example5() {
+    val result = listOf("a", "bbb", "cc").sortedBy { it.length }
+    result == listOf("a", "cc", "bbb")
+}
+
+fun Shop.getCustomersSortedByNumberOfOrders(): List<Customer> {
+    return this.customers.sortedBy { it.orders.size }
+}