+++ Others/Kotlin/kotlin-koans/src/ii_collections/n19Sum.kt
@@ -1,0 +1,10 @@
+package ii_collections
+
+fun example6() {
+    listOf(1, 3).sum() == 4
+    listOf("a", "b", "cc").sumBy { it.length } == 4
+}
+
+fun Customer.getTotalOrderPrice(): Double {
+    return this.orders.flatMap { it.products }.sumByDouble { it.price }
+}