+++ Others/Kotlin/kotlin-koans/src/ii_collections/n23CompoundTasks.kt
@@ -1,0 +1,13 @@
+package ii_collections
+
+fun Shop.getCustomersWhoOrderedProduct(product: Product): Set<Customer> {
+    return this.customers.filter { it.orderedProducts.contains(product) }.toSet()
+}
+
+fun Customer.getMostExpensiveDeliveredProduct(): Product? {
+    return this.orders.filter { it.isDelivered }.flatMap { it.products }.maxBy { it.price }
+}
+
+fun Shop.getNumberOfTimesProductWasOrdered(product: Product): Int {
+    return this.customers.flatMap { it.orders.flatMap { it.products } }.filter { it == product }.count()
+}