+++ Others/Kotlin/kotlin-koans/src/v_builders/n37StringAndMapBuilders.kt
@@ -1,0 +1,35 @@
+package v_builders
+
+import util.TODO
+import java.util.*
+
+fun buildStringExample(): String {
+    fun buildString(build: StringBuilder.() -> Unit): String {
+        val stringBuilder = StringBuilder()
+        stringBuilder.build()
+        return stringBuilder.toString()
+    }
+
+    return buildString {
+        this.append("Numbers: ")
+        for (i in 1..10) {
+            // 'this' can be omitted
+            append(i)
+        }
+    }
+}
+
+fun buildMap(build: HashMap<Int, String>.() -> Unit): HashMap<Int, String> {
+    val result = HashMap<Int, String>()
+    result.build()
+    return result
+}
+
+fun task37(): Map<Int, String> {
+    return buildMap {
+        put(0, "0")
+        for (i in 1..10) {
+            put(i, "$i")
+        }
+    }
+}